dissector: add a lot of enum values
Replace the tables of enums with somewhat-generated ones.  Also split
each strin table in two: one (_DESC_MAP) used for the dissection tree
(which sometimes include an English description), and one (_STR_MAP)
that's simply the string name of each constant (which are used e.g. in
the Info column for brevity).

Signed-off-by: Alex Badea <alex.badea@keysight.com>